% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_em.R
\name{run_em}
\alias{run_em}
\title{Estimate parameters for a multivariate time series}
\usage{
run_em(dir)
}
\arguments{
\item{dir}{A file path specifying
where you want the simulation results to be saved.
The path can either be a full directory or relative to
your current working directory, where the default is \code{getwd()}.}
}
\value{
todo: document this further.
}
\description{
todo: details go here
}
\examples{

}
\seealso{
Other dir input: \code{\link{run_om}},
  \code{\link{run_sim}}
}
\author{
Kelli Faye Johnson, \email{kelli.johnson@noaa.gov}
}
\concept{dir input}
