% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_MARSS.R
\name{run_MARSS}
\alias{run_MARSS}
\title{Calculate estimates from a MARSS model}
\usage{
run_MARSS(data, MARSScntliterations = 500, MARSSbootn = 100, B, Q,
  U = "zero", Z, R, A = NULL)
}
\arguments{
\item{data}{A \code{data.frame} with two columns,
where the first column is the attribute and the second column is the
indicator, or the dependent variable. The columns need not be named.}

\item{MARSScntliterations}{An integer value specifying the maximum number of
iterations to run in the estimation model.}

\item{MARSSbootn}{An integer value specifying the number of bootstrap
iterations to run generating the confidence intervals.}

\item{B}{Specifications for interactions in the estimation method.}

\item{Q}{Specifications for process error in the estimation method.}

\item{U}{Specifications for intercepts in the estimation method.}

\item{Z}{Specifications for observations to hidden states
in the estimation method.}

\item{R}{Specifications for observation error in the estimation method.}

\item{A}{Specifications for scaling matrix in the estimation method.}
}
\value{
A list object is returned with the model results and names of
the attributes and indicators, along with the original data used to run
the model and a status indicator providing information on if the model was
ran (i.e., "good") or not (i.e., "bad").
}
\description{
Calculate estimates from a MARSS model
}
\details{
Run a multivariate-autoregressive state-space (MARSS) model.
The model structure assumes that there are two time series available,
where the attribute affects the indicator. The magnitude of this affect
will be estimated using the $B$ matrix in the off-diagonal lower element.
The MARSS model is ran using \code{\link[MARSS]{MARSS}}.
Warning messages from \code{\link[MARSS]{MARSS}} are suppressed using
\code{sink} and \code{\link[base]{tryCatch}}.
}
\author{
Kelli Faye Johnson
}
