% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_om.R
\name{run_om}
\alias{run_om}
\title{Simulate multivariate time-series data}
\usage{
run_om(ompars = NULL, dir = getwd())
}
\arguments{
\item{ompars}{A dataframe of parameter values for the OM.
Each row specifies a single scenario.}

\item{dir}{A file path specifying
where you want the simulation results to be saved.
The path can either be a full directory or relative to
your current working directory, where the default is \code{getwd()}.}
}
\value{
todo: document this further.
}
\description{
todo: details go here
}
\examples{

}
\seealso{
Other dir input: \code{\link{run_em}},
  \code{\link{run_sim}}
}
\author{
Kelli Faye Johnson, \email{kelli.johnson@noaa.gov}
}
\concept{dir input}
