% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sim.R
\name{run_sim}
\alias{run_sim}
\title{Simulate data and estimate parameters for time series}
\usage{
run_sim(dir = getwd())
}
\arguments{
\item{dir}{A file path specifying
where you want the simulation results to be saved.
The path can either be a full directory or relative to
your current working directory, where the default is \code{getwd()}.}
}
\value{
todo: document this further.
}
\description{
todo: details go here
}
\examples{

}
\seealso{
Other dir input: \code{\link{run_em}}, \code{\link{run_om}}
}
\author{
Kelli Faye Johnson, \email{kelli.johnson@noaa.gov}
}
\concept{dir input}
