% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvvarss_simTVVAR.R
\name{scale2var}
\alias{scale2var}
\title{Scale the variance of a multivariate time series}
\usage{
scale2var(B, varinf = NULL, vare = NULL)
}
\arguments{
\item{B}{A square matrix of interactions.}

\item{varinf}{A square matrix of overall variance on
the diagonal and the covariance in the off diagonal
that will be used to solve for the process error of the
time series given the stationary distribution.}

\item{vare}{A square matrix of process errors that
will be used to solve for the stationary variance
of the time series if you had an infinite sample size.}
}
\description{
Scale the variance of a multivariate time series
}
\details{
From Ives \emph{et al}. (2003) Equation 17.
The variance matrix will be used to solve for the other
matrix, where both entries cannot be \code{NULL}. One always needs
to specify the \code{\strong{B}} matrix.
}
\examples{
scale2var(varinf = matrix(c(1, 0, 0, 1), nrow = 2, ncol = 2),
  B = matrix(c(-0.9, 0.1, 0.1, 0.9), nrow = 2))
}
\author{
Kelli Faye Johnson
}
